requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.3
Pillow==10.1.0
python-dateutil==2.8.2